pyproject.toml
[tool.poetry]
name = "sudoku"
version = "0.1.0"
description = "a brute force sudoku solver"
authors = ["Tim den Dulk <timdendulk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.0"
numpy = "^1.21.0"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
pre-commit = "^2.13.0"
flake8 = "^3.9.2"
isort = "^5.9.1"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
